Move WheelBuilder and friends to a dedicated module
